+++ src/Type/InputRichMessage.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Phptg\BotApi\Type;
 
+use Phptg\BotApi\FileCollector;
+
 /**
  * @see https://core.telegram.org/bots/api#inputrichmessage
  *
@@ -11,19 +13,30 @@ namespace Phptg\BotApi\Type;
  */
 final readonly class InputRichMessage
 {
+    /**
+     * @param InputRichMessageMedia[]|null $media
+     */
     public function __construct(
         public ?string $html = null,
         public ?string $markdown = null,
         public ?true $isRtl = null,
         public ?true $skipEntityDetection = null,
+        public ?array $media = null,
     ) {}
 
-    public function toRequestArray(): array
+    /**
+     * @return array<string, mixed>
+     */
+    public function toRequestArray(?FileCollector $fileCollector = null): array
     {
         return array_filter(
             [
                 'html' => $this->html,
                 'markdown' => $this->markdown,
+                'media' => $this->media === null ? null : array_map(
+                    static fn(InputRichMessageMedia $media) => $media->toRequestArray($fileCollector),
+                    $this->media,
+                ),
                 'is_rtl' => $this->isRtl,
                 'skip_entity_detection' => $this->skipEntityDetection,
             ],